implemented basic handle bars setup. there is a bookshelf route in the server file (needs to be moved afterwards) and the data is bring pulled from the DB. added new folders, moved files as well.
