+++ server/controllers/routes/api/bookRoutes.js
@@ -1,0 +1,5 @@
+const router = require('express').Router()
+const sequelize = require('../../../config/connection')
+const { User, Book } = require('../../../models')
+
+module.exports = router
