+++ server/package.json
@@ -13,6 +13,7 @@
   "dependencies": {
     "dotenv": "^16.0.3",
     "express": "^4.18.2",
+    "express-handlebars": "^6.0.6",
     "handlebars": "^4.7.7",
     "mysql2": "^2.3.3",
     "sequelize": "^6.26.0"
