--- server/routes/api/bookRoutes.js
@@ -1,7 +1,0 @@
-const router = require('express').Router()
-const sequelize = require('../../config/connection')
-const { User, Book } = require('../../models')
-
-// routes here
-
-module.exports = router
