+++ server/views/layouts/main.handlebars
@@ -1,0 +1,28 @@
+<!DOCTYPE html>
+<html lang="en">
+<head>
+  <meta charset="UTF-8">
+  <meta http-equiv="X-UA-Compatible" content="IE=edge">
+  <meta name="viewport" content="width=device-width, initial-scale=1.0">
+  <title>My Bookshelf</title>
+
+  <link rel="stylesheet"  href="/css/reset.css">
+  <link rel="stylesheet"  href="/css/styles.css">
+</head>
+<body>
+  <nav>
+    <a href="/">Home</a>
+    <a href="/about">About</a>
+  </nav>
+
+  <header>
+    <h1>Welcome back to your bookshelf</h1>
+  </header>
+  
+  <main>
+    <div class="bookshelf">
+      {{{body}}}
+    </div>
+  </main>
+</body>
+</html>
